+++ packages/expo-modules-core/CHANGELOG.md
@@ -21,6 +21,7 @@
 ### 🐛 Bug fixes
 
 - [iOS] Add async/await overload for StaticAsyncFunction ([#44471](https://github.com/expo/expo/pull/44471) by [@Wenszel](https://github.com/Wenszel))
+- [android] Fix permission status incorrectly showing `denied/canAskAgain=false` after user resets permission to "Ask every time" in system Settings. ([#44180](https://github.com/expo/expo/pull/44180) by [@Nedunchezhiyan-M](https://github.com/Nedunchezhiyan-M))
 - [iOS] Fixed looking up the `EXConstants.bundle` from the main bundle to allow running with precompiled XCFrameworks. ([#44551](https://github.com/expo/expo/pull/44551) by [@chrfalch](https://github.com/chrfalch))
 - [iOS] Fixed Constants.expoConfig returning null due to optional value wrapping in ConstantsProvider. ([#44550](https://github.com/expo/expo/pull/44550) by [@chrfalch](https://github.com/chrfalch))
 - [iOS] Fixed view lookup in bridgeless mode by using `ExpoHostWrapper` instead of `reactBridge.uiManager`. ([#44375](https://github.com/expo/expo/pull/44375) by [@vonovak](https://github.com/vonovak))